// src/components/oauth/_generated/server.ts
type Env = {
  readonly OAUTH_CLIENT_ID: string;
  readonly OAUTH_CLIENT_SECRET: string;
  readonly OAUTH_REDIRECT_URI: string | undefined;
  readonly PROVIDER: "google" | "github";
  readonly SITE_URL: string;
};
// ...
export const env: Env = process.env as unknown as Env;